:package: dependencies: Add black as a development dependency
This is a selfish change. I use a Vim plugin for the P.S.F. "Black"
Python tool. Because I am using Pipenv which creates Python virtual
environments, the Vim plugin breaks and I get an annoying error every
time I use Vim inside the virtual environment.

So, this is a subtly selfish change in that it makes it more convenient
for me to edit files on this project. We don't have a practical need for
Black right now but I am not equipped enough to spend time fixing this
issue right now.

Signed-off-by: Justin W. Flory <git@jwf.io>