:memo: Issue templates: Create a "New Story" template (closes #1)
This commit moves the repository over to use multiple issue templates
simultaneously. One is more customized for accepting new contributions
as "stories", which are how we describe the practical applications of
the Principles.

I hope this issue template facilitates easier contributions of new
stories. For now, I think this helps enable us to build this work
further.

Closes #1.

Signed-off-by: Justin W. Flory <git@jwf.io>